move WithSIGTERMCancel from Proby pkg to cobra command
